dashboard/requirements.txt
streamlit==0.78.0
pandas==1.0.1
numpy==1.18.1
Pillow==7.1.2
requests==2.22.0
azure-storage-blob
azure-cosmosdb-table
spacy==2.2.1